refactor(config): move meta related config into own file
